Evolution Arch, API and new crawler

+++ migrations/versions/1b750a389c22_remove_email_notification_column.py
@@ -10,13 +10,16 @@ Create Date: 2015-02-25 23:01:07.253429
 revision = '1b750a389c22'
 down_revision = '48f561c0ce6'
 
+import conf
 from alembic import op
 import sqlalchemy as sa
 
 
 def upgrade():
-    op.drop_column('feed', 'email_notification')
+    if 'sqlite' not in conf.SQLALCHEMY_DATABASE_URI:
+        op.drop_column('feed', 'email_notification')
 
 
 def downgrade():
-    op.add_column('feed', sa.Column('email_notification', sa.Boolean(), default=False))
+    op.add_column('feed', sa.Column('email_notification', sa.Boolean(),
+                  default=False))
